Bugfix 0309 (#503)
* Sup product info (#471)

* update

* update

* update

* add default sku 43A65H in author editor

* add sku

---------


* header (#472)

* header

* eslint

---------


* Bug 0309 (#475)

* header

* eslint

* css

---------


* 处理浮点数问题和样式问题

* fix style

* fix legal mobile style

* Bugfix 0309 (#476)

* 处理浮点数问题和样式问题

* fix style

* fix legal mobile style

* fix bugs (#477)

* fix bugs

* fix bugs

* fix bugs

* update title font-weight for region selection block (#478)


* reset merge conflict file

---------

+++ blocks/media-carousel/media-carousel.js
@@ -21,7 +21,7 @@ function bindEvent(block, type = 'normal') {
   if (block.classList.contains('bottom-center-style')) {
     maxWidth = block.offsetWidth;
   }
-  const gap = parseInt(window.getComputedStyle(block.querySelector('.media-carousel-track')).gap, 10) || 0;
+  const gap = parseFloat(window.getComputedStyle(block.querySelector('.media-carousel-track')).gap, 10) || 0;
   const CONFIG = {
     itemWidth: getChildSlideWidth(block),
     gap,
